+++ src/OnlineService.Data/Repositories/EntityBaseRepository.cs
@@ -36,11 +36,11 @@ namespace OnlineService.Data.Repositories
             return await query.ToListAsync();
         }
 
-        public T GetSingle(int id)
+        public T GetSingle(long id)
         {
             return _context.Set<T>().FirstOrDefault(x => x.Id == id);
         }
-        public async Task<T> GetSingleAsync(int id)
+        public async Task<T> GetSingleAsync(long id)
         {
             return await _context.Set<T>().FirstOrDefaultAsync(x => x.Id == id);
         }
